Changed componentFile to component.
Changed the test on componentFile to look for undefinedness or nullness
of component role. Beware that ListModels are very very very stupid.
The first element defines the roles for all the elements contained in the ListModel.
So, if component property is defined in the first element
and not defined in a following element, the property will be undefined on the element without it.
If the component property is not defined on the first element,
then all elements in the ListModel don't define it!!!
So, if you don't have a component property for the first element,
define a first element in the ListModel that contains
a key-value pair with key component and value undefined.