Changed translations management.
Refactored translations management inside *Names.qml files.
This is because Room system needs all translations.
To load them, all init functions are changed to inject translation to a container passed in.
This is useful because we can now inject all translation in a Names.qml component.
This component can be used in Room system to make all translations available.