Paginator list height computation.
In cases where the model has several items that may potentially
span over several pages, but the filtered items are contained in
only one page, we had a problem on height computation.
Height must consider that paginator is not visible and so must
be updated to take into account this.
Before the fix, height was not updated correctly if paginator
height was initially added.