Created a ControlLeftRight component.
This component is meant to be used when there is a fixed list of
choices and the user has to pick one of them.

The buttons contain images of left and right arrows.

It is similar to ControlUpDown (where arrows are in the up and
down directions), but uses 3 state buttons.

Signed-off-by: Roberto Gianassi <roberto@develer.com>