Bug fix on setting functions for ThermalControlUnit.
Now, only valid values are set.
The downside of actual implementation is that control blocks in
some situations.
For example, if you set a date of 30/1/2012 and try to increment
the month to February, month remains to January.
This implementation avoids ugly graphical glitches, but it has
to be revisited.

Signed-off-by: Roberto Gianassi <roberto@develer.com>