Removed the count property from MediaDataModel.
This property was present solely for historical purposes (it was used in
antintrusion system), but it was implemented asymmetrically, in that the
signal countChanged() wasn't emitted when adding elements.

Also, all the filters attached to the DataModel weren't correctly updated
on add or remove operations, which led to inconsistent results, for example
two filters on the same DataModel could report different objects count.

This patch correctly updates the filters on DataModel change, removes the
count property on MediaDataModel and updates the QML files to use an
ObjectModel instead of ObjectDataModel directly.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>