Introducing new isSelected property in MenuItem component.
The pressed state defined with a "when" transition was switching
to default state transitioning to false.
This is according to documentation about "when" property, but it
causes the lost of the selected state (even when the menu item
element remained selected).
The introduction of isSelected property permits to define a
when property for the selected state, too.
In this way, the selected state is always recorded and never
lost when transitioning through the pressed state.
All menu item are still to be updated to take advantage of
the new isSelected property.

Signed-off-by: Roberto Gianassi <roberto@develer.com>