First implementation of a ChoiceList class.
This class is needed to manage a set of choices taken from a bigger
set.
For example, a fancoil may use only some of all the possible values
so we need to manage such a situation.
This is a first skeleton for the ChoiceClass applied to Split Modes
(for tests).

Signed-off-by: Roberto Gianassi <roberto@develer.com>