Implementation of a denser grid.
Implementation of a moving sidebar around quicklinks to make them
more flexible.
The grid is denser, so a finer position is now possible.
Grid position and dimension computation is done to avoid that
quicklinks go out of grid. This avoids overlapping with other
elements or escaping out of screen.

Still to be implemented in rooms, awaiting for clarifications
on specifics.

Signed-off-by: Roberto Gianassi <roberto@develer.com>