Created a component to manage buttons/functions during a call.
This compoonent manages buttons and functions during a call.
When a call is ringing, it shows two buttons: one to pick up the call
and the other to reject it.
When the call is answered, it shows a button to terminate the call.
No text and/or image are shown. There will be an advanced control
for this.

Signed-off-by: Roberto Gianassi <roberto@develer.com>