Bug fix for spurious VDE messages.
In VDE system, some spurious messages are delivered to devices.
To avoid malfunctionings, a call active flag has been implemented.
When the flag is not set, all messages are ignored except for
ringtone and start call ones.
When the flag is set, all messages are processed has expected, but
only by the the interested object.
Please note, Intercom calls may be initiated with a direct call to
the device, so the flag has to be set there, too.

Signed-off-by: Roberto Gianassi <roberto@develer.com>