Added the real path (as QString) to FileObject class.
Introduce a logicalPath property which holds the full path of the FileObject
as a VariantList. In case of file system objects, the logical path is the same
as the path; for upnp files, they differ because usually upnp paths are URLs.

The logical path can be used when there's the need to know the path of the
file which is somewhat user friendly; for example, upnp path usually are like
http://mediaserverip/custom/path while the logical path in this case would
be ["mediaserver name", "path", "to", "file"].

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>