Play files correctly when selecting a different source from the current one.
When trying to play a different source from the current one, the current
source didn't get a notification that the playlist should change, so when the
player was stopped by another object it requested the next element in the
playlist.
This led to segfaults or infinite loops (as two different objects where stepping
on each other toes).

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>