Bug fix for static const int properties.
When a const is defined as static const int inside a class, it can be
compiled inline.
In such a case its address becomes unavailable, so if it is accessed
with a reference to int parameter, it causes a link error (undefined
reference).
We switched to use anonymous enums inside classes instead of static
const int to avoid such nasty problems.

Signed-off-by: Roberto Gianassi <roberto@develer.com>