Simplify Room.qml by removing the possibility to close the menu.
There was a clickable area that covered all the vertical space between
the toolbar and the room list at the bottom (but it didn't include the dark
background).
This was wrong in many ways:
 * covered the systems button, so the user had to press twice when a menu
was opened to return to the rooms starting page;
 * it was too small to click on the touch screen: basically the only clickable
part was the navigation bar;
 * UX wise, it's better to let the dark background be clickable to close the
menu.

This greatly simplifies the code and allows the removal of some functions
in RoomView.qml too.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>