Rework QuickLink z order in Profile page.
In some cases, when selecting a quicklink other quicklinks were still
selectable, ie. they were above the dark background.
This was due to the z order set a creation.

This patch fixes the bug and simplifies the z order for quicklinks
while we wait for instructions on the desired behaviour.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>