Bug fix for loadComponent not working on first created component.
The loadComponent is called in response to signal loadColumn.
The connection to this signal is operated inside the same loadComponent
function.
If not done properly, it causes some missed loadColumn signals and
weird observed behavior.
Now, as soon as we verify that the newly created itemObject is
"fine", we connect all needed signals and only then we proceed to
further elaborations.
In this way no signal is missed.

Signed-off-by: Roberto Gianassi <roberto@develer.com>