Move GStreamer media player plugin into its own directory.
There was an infinite recursion bug on the hardware. The algorithm to
load the plugin searched the application directory and tried to
instantiate every file as a plugin until it found the GStreamer plugin.
Unfortunately, the ARM version has a link to libbtobjects plugin in the
same directory, so the BtObjects plugin was loaded recursively.

Fixed by moving the GStreamer plugin into its own directory.