Ticket #17305: remove QmlApplicationViewer destruction/recreation.
Destroying and recreating the viewer causes some random corruption deep
into QtScript bowels.  To work around the issue avoid destruction/recreation
(given the way the GUI works now, it's not required anymore).

Leaving BootManager as-is for now: with some refactoring it could be used
for the browser in a separate process.

Signed-off-by: Mattia Barbon <mattia@develer.com>