Parse the general amplifier object.
Since we don't have an ambient for the general, we build one. This way we can
treat it like any other ambient in the GUI; it also naturally handles the concept
of source, which normal amplifiers don't have.

The General amplifier is created as an AmplifierGroup so that the graphical
representation is correct, since there's no concept of active or volume
level.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>