Ticket #18266: better error handling.
WebView.onLoadFailed does not have a way to differentiate errors while
loading main page content and embedded objects.

The correct way would be overriding QWebPage::extension(), but this is
not possible when using QML WebView.

The implementation uses a wrapper around QNetworkReply that provides
default error content when the error response is empty.

Signed-off-by: Mattia Barbon <mattia@develer.com>