Add LoaderWithProps.
Provides Loader-like functionality and provides creation time properties.
It should act as a drop-in replacement for Loader; only tested for popup
elements in BasePage for now.

Provinding properties at build time avoids many runtime errors associated
with Loader, where the created component requires some properties to be
set (impossible with standard Loader).
The only drawback is that the component must be created using the function
setComponent() instead of directly assigning to the item property.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>