Make sure we reboot every time we are in 'reboot' state.
After commit 7d251625cb2ce3f73550cc5146ea25cc4c0104b7, there were
some cases where the UI was in 'reboot' state (user input blocked, loading
indicator) but the UI was not rebooted properly.
Such cases arise when we try to set a property with the same value as the
current one: in this case the property is not modified and the change to
conf.xml is not scheduled.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>