Better implementation of UI update when closing menus.
Commit 0372f0772ba0732ef32cdd609a2d4a0d073021cf introduced a
heuristic for updating the UI. However this introduced a new bug: in home
and room settings it's not possible anymore to switch between stock images
and mediaserver menus.
In those cases, the implementation put 2 consecutive UPDATE_UI operations.

Fixed by updating the UI only when the size of the columns to be closed is
wider than the visible area in mainContainer.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>