Fix bug with short texts in navigation bar, introduced with commit 1fb8d77e
The text item is transformed relative to the original position of the text, not
relative to the size of the item.
This is fine until the size of the element is more than the size of the text, as
the commit above did. With this configuration, centering is done relative to
the Item, while the rotation is done relative to the text, resulting in text
that was drawn outside the screen.
Fixed by using an Item as a positioner, then drawing the text inside it.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>