Fix #18762, the folder icon returns to the previous visited folder.
UPnP sources: avoid closing the menu column, since this conflicts with the
feature. In fact, the playing status changes as soon as we start any playback
via UPnP.
This is not a big problem because  the only time we see the page we the
browser already opened is when we come back from a player. In the other
cases the menu is closed because the page is created from scratch.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>