Avoiding spurious offset displacements.
If a user double clicks on an arrow on a path view, the second
click arrives when the first is still "executing".
The offset is still not integer. Adding 1 at that point
causes an update to offset value to a not rounded value like
2.3343333... .
Testing for roundness avoids this glitch.

Signed-off-by: Roberto Gianassi <roberto@develer.com>