Air conditioning: allow pagination in program list.
This requires an ObjectDataModel for program lists in basic and advanced
splits.
The rest of the patch basically adapts the rest of the world to this new
interface. In particular, the tests now need to create the program list on the
heap because the ObjectDataModel takes ownership of the pointer passed in.

I've tried hard to keep the change count low and not to break the
tests and the GUI code; as a result, there are points in the code that need
other refactorings.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>