Bug #19749, freeze on multiple scenario activation popups.
Fix synchronization issues in the advanced scenario popup logic. 
If a user presses too quickly, there may be two or more events that arrive
on the page while we are already showing a popup for scenarios. We must
correctly queue such events.
I've not used dismissTimer because during the onTriggered slot the running
property is still true, and it's set to false immediately after; any logic that
restarts the timer in the onTriggered slot is doomed to failure.
Also, create a proper sync signal in BasePage to know when a popup is really
dismissed.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>