Bug #20067, AUX source control not visible.
If we switch to AUX after a reboot with a radio source active, the source
control for AUX is hidden.
Here is what happens:
* the RDS radio control is loaded into the AnimatedLoader
* we select the AUX source: the new source control is loaded
* the RDS radio is switched off from bus: we have an empty source, which is
loaded into the loader
* the AUX source is turned on from bus: bug! We try to set the the AUX source
control in the loader, but it is already present (only hidden), so we don't do
anything.

Fixed by properly destroying the original component in
AnimatedLoader.destroyComponent()

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>