Bug #20335, shut down video during an intercom call.
This code was taken from MediaPlayer::stop(), but in that case there was
an active flag to be checked. In our case, we only have 1 instance of the
gstreamer player, so we don't need the check above.
This code now matches the code path taken in MediaPlayer when the player
is active.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>