Bug #20256. New implementation for card and background paths.
Paths for cards and backgrounds are now relative in layout.xml.
Their values starts with images/ or custom_images/.
In the former case, they reside in extra/1 so, when building the
absolute path, use extra/1.
In the latter case, they reside in extra/12 and we have to use
that path to build the absolute one.
On PCs, we have also to consider that custom images are saved in
a folder which is in the project folder, in bin/x86/12 subfolder.

This commit uses the new implementation for the home page
background. Implementations for profiles/rooms cards/backgrounds
will follow.

Signed-off-by: Roberto Gianassi <roberto@develer.com>