Bug #20820. Taking a snapshot of config value in ctor.
During object construction, we need to take a snapshot of config
values because set calls logic may overwrite member values.

SetSkin and SetHomeBgImage logic works if we change only one
value at a time. This is because the skin and background values
are strictly related between them. In fact, the home page has
2 default backgrounds that change when skin changes.
Every time we change the skin and the home background is not
customized, we have to update the img tag. So each one of this
method changes its value considering the other constant.

In object construction, we need to set both at the same time
reading them from config data. In this case, we take a snapshot
of config values. Then we set the skin (which causes the home
background to be updated to one of the default values) and then
we set the background (just in case it was customized).
In this way, we rest assured that correct data is loaded in the
constructor.

Signed-off-by: Roberto Gianassi <roberto@develer.com>