Rename autoTurnOff property into timingEnabled and improve Light interface.
The autoTurnOff property was used in two different use cases (custom timing
and fixed timing) and also the frame sending code was relying on it to be set
even though it was not strictly necessary. This complicated client code.

Now custom timing and fixed timing are controlled by different properties.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>