Made browserItem the BasePage.
When using the Scale transform, the rescaling is applied only
graphically. A rescaled item still have its original size,
but is rendered scaled on screen.
If the browser main page is a BasePage, this causes some
weird behaviors. The browser page has wrong dimensions,
popups loaded by this page are not positioned properly
and so on.
Making the browserItem the BasePage component solves the issue
because the browser page itself is created with the reference
dimension of 1024x600. At this point, all components created
as children are referring to this reference dimension and
so all rescaling, positioning, ... operations are applied properly.

Signed-off-by: Roberto Gianassi <roberto@develer.com>