Bug #23998, refactor upnp media server usage.
Allow the user to stop the running media server directly from the popup.

Also cover the following case:
the user selects media server as a source in an ambient, switches to multimedia
and starts a media server stream, returns to the previous ambient and browses
the songs in the media server.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>