Fix visualization of files in the root directory of local mounts.
The problem was that during exitDirectory() we cleared both the current_path
and page_position. The first must not be cleared for local files (and it's useless
for upnp models since if we are at root it's already empty), the second must
not be reset in any case (because it contains the range for the root directory).

With this patch we also get rid of the serverList property, which is specific of
the upnp implementation.

Signed-off-by: Luca Ottaviano <lottaviano@develer.com>